+++ DirectX12Lib/LinearAllocator.h
@@ -23,7 +23,7 @@ enum ELinearAllocatorType
 struct FAllocation
 {
 	ID3D12Resource* D3d12Resource;
-	uint64_t Offset;
+	size_t Offset;
 	void* CPU;
 	D3D12_GPU_VIRTUAL_ADDRESS GpuAddress;
 };
@@ -47,7 +47,7 @@ private:
 	
 private:
 	uint64_t m_FenceValue;
-	uint32_t m_PageSize;
+	size_t m_PageSize;
 	ComPtr<ID3D12Resource> m_d3d12Resource;
 	void* m_CpuAddress;
 	D3D12_GPU_VIRTUAL_ADDRESS GpuAddress;
@@ -59,21 +59,21 @@ class LinearAllocator
 public:
 	LinearAllocator(ELinearAllocatorType Type);
 
-	FAllocation Allocate(uint32_t SizeInBytes, uint32_t Alignment = DEFAULT_ALIGN);
+	FAllocation Allocate(size_t SizeInBytes, size_t Alignment = DEFAULT_ALIGN);
 
 	void CleanupUsedPages(uint64_t FenceID);
 	void Destroy();
 
 private:
 	LinearAllocationPage* AllocateLargePage(uint32_t SizeInBytes);
-	LinearAllocationPage* RequestPage(uint32_t SizeInBytes);
-	LinearAllocationPage* CreateNewPage(uint32_t PageSize);
+	LinearAllocationPage* RequestPage(size_t SizeInBytes);
+	LinearAllocationPage* CreateNewPage(size_t PageSize);
 
 	using PagePool = std::deque<LinearAllocationPage* >;
 
 	ELinearAllocatorType m_AllocatorType;
-	uint32_t m_PageSize;
-	uint32_t m_CurrentOffset = 0;
+	size_t m_PageSize;
+	size_t m_CurrentOffset = 0;
 
 	PagePool m_UsingPages;
 	PagePool m_RetiredPages;
